+++ CHANGELOG.md
@@ -8,9 +8,11 @@
 ### 新特性
 * 【core】        MapUtil增加newConcurrentHashMap（pr#538@Github）
 * 【core】        增加StopWatch（issuepr#539@Github）
+* 【core】        增加ZipUtil.listFiles（issuepr#541@Github）
 
 ### Bug修复
 * 【core】        修复DateUtil.endOfYear计算错误问题（issuepr#540@Github）
+* 【core】        修复FileUtil.listFileNames在jar中匹配问题，增加（issuepr#541@Github）
 
 -------------------------------------------------------------------------------------------------------------